requirements.txt
asgiref==3.4.1
Django==3.2.6
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-taggit==1.5.1
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
